+++ connect.js
@@ -6,6 +6,7 @@ import { withConnect } from './ConnectContext';
 
 import OkapiResource from './OkapiResource';
 import RESTResource from './RESTResource';
+import { initialResourceState } from './RESTResource/reducer';
 import LocalResource from './LocalResource';
 import { mutationEpics, refreshEpic } from './epics';
 
@@ -120,7 +121,8 @@ const wrap = (Wrapped, module, epics, logger, options = {}) => {
       });
     }
 
-    UNSAFE_componentWillReceiveProps(nextProps) { // eslint-disable-line react/no-deprecated
+    // eslint-disable-next-line camelcase, react/no-deprecated
+    UNSAFE_componentWillReceiveProps(nextProps) {
       // this.logger.log('connect', `in componentWillReceiveProps for ${Wrapped.name}: nextProps.location=`, nextProps.location, 'this.props.location=', this.props.location);
       if (this.componentShouldRefreshRemote(nextProps)) {
         this.props.refreshRemote({ ...nextProps });
@@ -135,6 +137,10 @@ const wrap = (Wrapped, module, epics, logger, options = {}) => {
 
           if (resource.shouldReset()) {
             resource.reset();
+          }
+
+          if (!resource.isVisible()) {
+            resource.cancelRequestsOnUnmout();
           }
         }
       });
@@ -178,10 +184,20 @@ const wrap = (Wrapped, module, epics, logger, options = {}) => {
     logger.log('connect-lifecycle', `mapState for <${Wrapped.name}>, resources =`, resources);
 
     const resourceData = {};
-    for (const r of resources) {
-      resourceData[r.name] = Object.freeze(_.get(state, [`${r.stateKey()}`], null));
-    }
 
+    for (const r of resources) {
+      let initState = _.get(state, r.stateKey());
+
+      if (!initState) {
+        if (r instanceof OkapiResource) {
+          initState = initialResourceState;
+        } else {
+          initState = r?.query?.initialValue !== undefined ? r.query.initialValue : {};
+        }
+      }
+
+      resourceData[r.name] = Object.freeze(initState);
+    }
     const newProps = { dataKey, resources: resourceData };
     // TODO Generalise this into a pass-through option on connectFor
     if (typeof state.okapi === 'object') newProps.okapi = state.okapi;