+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Change history for stripes-connect
 
-## 5.7.0 (IN PROGRESS)
+## 6.2.0 (IN PROGRESS)
+
+* Added the ability for elements of a manifest 'headers' property to be interpreted, and also to be a function in addition to an object. Refs STCON-121
+
+## [6.1.0](https://github.com/folio-org/stripes-connect/tree/v6.1.0) (2021-02-25)
+[Full Changelog](https://github.com/folio-org/stripes-connect/compare/v6.0.0...v6.1.0)
+
+* Added additional check to not trigger a fetch when params is null, refs STCON-115
+* Perform substitutions on perRequest option, refs STCON-117
+
+## [6.0.0](https://github.com/folio-org/stripes-connect/tree/v6.0.0) (2020-10-06)
+[Full Changelog](https://github.com/folio-org/stripes-connect/compare/v5.6.1...v6.0.0)
+
+* Init resources correctly so their state is immediately available for inspection.
+* Add ability to cancel pending requests. Revs STCON-104.
+* Remove unnecessary use of isomorphic-fetch
+* Allow mutators to configure `throwErrors` option. STCON-112.
+
+## [5.6.1](https://github.com/folio-org/stripes-connect/tree/v5.6.1) (2020-06-08)
+[Full Changelog](https://github.com/folio-org/stripes-connect/compare/v5.6.0...v5.6.1)
 
 * Use `UNSAFE_componentWillReceiveProps` for a quieter console. Refs STCON-70.